Add EIC_gen.py script, Test Files, and requirements
EIC_gen hosts the EICgen class, which will generate a single EIC_CSV
compiled based on the input targets. The approach is to use the parse_mzXML
function from mzXML_to_csv.py to generate mzCSV files, and then to load
the generated file data into memory before proceeding to filter the data
based on inputs from the feature file and program arguments. Note this
approach may not work well with a large number of mzXML files, and
in that case we may want to consider a chunking approach.